requirements.txt
greenlet==3.2.3
numpy==2.3.2
pandas==2.3.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.42
typing_extensions==4.14.1
tzdata==2025.2
